refactor Fission.Storage.IPFS to folder + style tweaks
