+++ library/Fission/Storage/IPFS/DAG.hs
@@ -1,0 +1,4 @@
+module Fission.Storage.IPFS.DAG
+  ( put ) where
+
+import Fission.Storage.IPFS.DAG.Put
